Deleted livestreams fixed

+++ bellify_bot/management/commands/check_for_deleted_livestreams.py
@@ -1,6 +1,7 @@
 from django.core.management import BaseCommand
 from django.utils.timezone import now
 from youtube.models import YouTubeChannel, YouTubeVideo
+from youtube.utils import check_if_video_is_exists
 
 
 class Command(BaseCommand):
@@ -9,12 +10,13 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         for channel in YouTubeChannel.objects.all():
             for livestream in channel.livestreams.all():
-                saved_livestream = YouTubeVideo.objects.filter(
-                    channel=channel, video_id__exact=livestream.livestream_id).first()
+                if (now() - livestream.ended_at).total_seconds() > 3600 and not livestream.is_checked_for_deleted:
+                    livestream.set_checked()
 
-                if livestream.ended_at:
-                    if saved_livestream and ((now() - saved_livestream.create_at).seconds > 3600) and saved_livestream.skipped_iterations > 0:
+                    saved_livestream = YouTubeVideo.objects.filter(
+                        channel=channel, video_id__exact=livestream.livestream_id).first()
+
+                    if not check_if_video_is_exists(livestream.livestream_url):
+                        livestream.delete()
+                        saved_livestream.delete() if saved_livestream else None
                         channel.increment_deleted_livestreams()
-                        saved_livestream.delete()
-                    elif (now() - livestream.ended_at).seconds > 3600 and not livestream.counted_as_deleted:
-                        livestream.count_as_deleted()
